test: pull logging setup out of setUpAll hook
The test framework uses zones to associates print statements with the
test or hook which contains the print statement. Setting up logging in
a hook caused all log messages to be associated with that hook.

Now logging is setup when the test binding is initialized and in the
root zone.