feat: use consistent database default directory for Flutter
With this change the default database directory for Flutter apps is calculated by taking `getApplicationSupportDirectory` from `package:path_provider` and appending the path segment `CouchbaseLite`.

To enable this feature the concept of an isolate context is introduced, which is created when initializing the primary isolate and must be provided when initializing secondary isoalates.

Closes #168