fix: use different `MContext` for every `Result` in `ProxyResultSet`
This fixes an issue where the results of a result set are not properly
decoded, because shared string from one result set are used in another.

This can happen if the memory used for one result set is released and
reused for another.

The cause is that results shared a `SharedString` through the common
context, even though they don't actually share strings.