feat!: Add `Database.saveBlob` and `Database.getBlob`
This change adds the methods `Database.saveBlob` and `Database.getBlob`.
They allow saving and retrieving `Blob`s, independently of `Document`s.

This is a breaking change because `Blob.fromStream` is now a factory
instead of a static method returning a `Future`. To get the previous
behavior, create a `Blob` with `Blob.fromStream` and save it with
`Database.saveBlob`.

This change also fixes `Blob.isBlob`.

`Blob.properties` now returns a `Map` that contains full `Blob`
metadata.

Closes #255
Closes #253