refactor: remove `_ProxyQueryEarlyFinalizer`
In normal operation proxy objects release their corresponding objects
on the remote side when they are GCed.

The intention of `_ProxyQueryEarlyFinalizer` was to ensure the relase of
the `Query` object on the remote side, when the `ProxyDatabase` closes
before the `ProxyQuery` has been GCed.
When the `ProxyDatabase` closes the `CblService` that potential still
holds onto the `Query` through a `ObjectRegistry`, is soon GCed anyway.
A `WorkerDatabase` tears down its `IsolateWorker` when it closes. Besides
for testing there currently is no other situation where `CblService` instances
are created.