refactor: remove `_ProxyQueryEarlyFinalizer` (#731)
In normal operation, proxy objects release their corresponding objects
on the remote side when they are GCed.

The intention of `_ProxyQueryEarlyFinalizer` was to ensure the release of
the `Query` object on the remote side, when the `ProxyDatabase` closes
before the `ProxyQuery` has been GCed.
When the `ProxyDatabase` closes, the `CblService` that potential still
holds on to the `Query` through a `ObjectRegistry`, is soon GCed anyway.
A `WorkerDatabase` tears down its `IsolateWorker` when it closes. Besides
for testing, there currently is no other situation where `CblService` instances
are created.

It turned out that the closing of `_ProxyQueryEarlyFinalizer` is run in parallel with 
closing `ListenerTokenRegistry` and `ProxyListenerToken`s of the same query. 
This caused the `Query` on the remote side to be released before the `ProxyListenerToken` 
had a change to remove its corresponding `ListenerToken` from the `Query`. 
This PR also fixes this issue.